cleanup codes and added bug report template
