+++ src/gdrive/gdrive.go
@@ -1,19 +1,14 @@
 package gdrive
 
 import (
-	"crypto/md5"
 	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
-	"path/filepath"
 	"regexp"
-	"strings"
-	"sync"
+	"net/http"
+	"path/filepath"
 
 	"github.com/KJHJason/Cultured-Downloader-CLI/request"
-	"github.com/KJHJason/Cultured-Downloader-CLI/spinner"
 	"github.com/KJHJason/Cultured-Downloader-CLI/utils"
 	"github.com/fatih/color"
 )
@@ -78,22 +73,6 @@ func (gdrive *GDrive) GDriveKeyIsValid() (bool, error) {
 	return res.StatusCode != 400, nil
 }
 
-type GDriveFile struct {
-	Kind        string `json:"kind"`
-	Id          string `json:"id"`
-	Name        string `json:"name"`
-	Size        string `json:"size"`
-	MimeType    string `json:"mimeType"`
-	Md5Checksum string `json:"md5Checksum"`
-}
-
-type GDriveFolder struct {
-	Kind             string       `json:"kind"`
-	IncompleteSearch bool         `json:"incompleteSearch"`
-	Files            []GDriveFile `json:"files"`
-	NextPageToken    string       `json:"nextPageToken"`
-}
-
 // Logs any failed GDrive API calls to the given log path
 func LogFailedGdriveAPICalls(res *http.Response, downloadPath string) {
 	requestUrl := res.Request.URL.String()
@@ -144,6 +123,22 @@ func LogFailedGdriveAPICalls(res *http.Response, downloadPath string) {
 		)
 		utils.LogError(err, "", false)
 	}
+}
+
+type GDriveFile struct {
+	Kind        string `json:"kind"`
+	Id          string `json:"id"`
+	Name        string `json:"name"`
+	Size        string `json:"size"`
+	MimeType    string `json:"mimeType"`
+	Md5Checksum string `json:"md5Checksum"`
+}
+
+type GDriveFolder struct {
+	Kind             string       `json:"kind"`
+	IncompleteSearch bool         `json:"incompleteSearch"`
+	Files            []GDriveFile `json:"files"`
+	NextPageToken    string       `json:"nextPageToken"`
 }
 
 // Returns the contents of the given GDrive folder
@@ -285,308 +280,4 @@ func (gdrive *GDrive) GetFileDetails(fileId, logPath string) (map[string]string,
 		"mimeType":    gdriveFile.MimeType,
 		"md5Checksum": gdriveFile.Md5Checksum,
 	}, nil
-}
-
-// Downloads the given GDrive file using GDrive API v3
-//
-// If the md5Checksum has a mismatch, the file will be overwritten and downloaded again
-func (gdrive *GDrive) DownloadFile(fileInfo map[string]string, filePath string) error {
-	if utils.PathExists(filePath) {
-		// check the md5 checksum and the file size
-		file, err := os.OpenFile(filePath, os.O_RDONLY, 0666)
-		if err != nil {
-			err = fmt.Errorf(
-				"gdrive error %d: failed to open file \"%s\", more info => %v",
-				utils.OS_ERROR,
-				filePath,
-				err,
-			)
-			return err
-		}
-		fileStatInfo, err := file.Stat()
-		if err != nil {
-			err = fmt.Errorf(
-				"gdrive error %d: failed to get file stat info of \"%s\", more info => %v",
-				utils.OS_ERROR,
-				filePath,
-				err,
-			)
-			return err
-		}
-		fileSize := fileStatInfo.Size()
-		if fmt.Sprintf("%d", fileSize) == fileInfo["size"] {
-			md5Checksum := md5.New()
-			_, err = io.Copy(md5Checksum, file)
-			file.Close()
-			if err == nil {
-				if fmt.Sprintf("%x", md5Checksum.Sum(nil)) == fileInfo["md5Checksum"] {
-					return nil
-				}
-			}
-		} else {
-			file.Close()
-		}
-	}
-
-	params := map[string]string{
-		"key":              gdrive.apiKey,
-		"alt":              "media", // to tell Google that we are downloading the file
-		"acknowledgeAbuse": "true",  // If the files are marked as abusive, download them anyway
-	}
-	url := fmt.Sprintf("%s/%s", gdrive.apiUrl, fileInfo["id"])
-	res, err := request.CallRequest("GET", url, gdrive.downloadTimeout, nil, nil, params, false)
-	if err != nil {
-		return err
-	}
-	defer res.Body.Close()
-	if res.StatusCode != 200 {
-		LogFailedGdriveAPICalls(res, filePath)
-		return nil
-	}
-
-	file, err := os.Create(filePath)
-	if err != nil {
-		err = fmt.Errorf(
-			"gdrive error %d: failed to create file \"%s\", more info => %v",
-			utils.OS_ERROR,
-			filePath,
-			err,
-		)
-		return err
-	}
-	defer file.Close()
-
-	_, err = io.Copy(file, res.Body)
-	if err != nil {
-		os.Remove(filePath)
-		err = fmt.Errorf(
-			"gdrive error %d: failed to download file \"%s\", more info => %v",
-			utils.DOWNLOAD_ERROR,
-			filePath,
-			err,
-		)
-		return err
-	}
-	return nil
-}
-
-// Downloads the multiple GDrive file in parallel using GDrive API v3
-func (gdrive *GDrive) DownloadMultipleFiles(files []map[string]string) {
-	var allowedForDownload, notAllowedForDownload []map[string]string
-	for _, file := range files {
-		if strings.Contains(file["mimeType"], "application/vnd.google-apps") {
-			notAllowedForDownload = append(notAllowedForDownload, file)
-		} else {
-			allowedForDownload = append(allowedForDownload, file)
-		}
-	}
-
-	if len(notAllowedForDownload) > 0 {
-		noticeMsg := "The following files are not allowed for download:\n"
-		for _, file := range notAllowedForDownload {
-			noticeMsg += fmt.Sprintf(
-				"Filename: %s (ID: %s, MIME Type: %s)\n",
-				file["name"], file["id"], file["mimeType"],
-			)
-		}
-		utils.LogError(nil, noticeMsg, false)
-	}
-
-	if len(allowedForDownload) > 0 {
-		maxConcurrency := gdrive.maxDownloadWorkers
-		if len(allowedForDownload) < maxConcurrency {
-			maxConcurrency = len(allowedForDownload)
-		}
-		var wg sync.WaitGroup
-		queue := make(chan struct{}, maxConcurrency)
-		errChan := make(chan map[string]string, len(allowedForDownload))
-
-		baseMsg := "Downloading GDrive files [%d/" + fmt.Sprintf("%d]...", len(allowedForDownload))
-		progress := spinner.New(
-			spinner.DL_SPINNER,
-			"fgHiYellow",
-			fmt.Sprintf(
-				baseMsg,
-				0,
-			),
-			fmt.Sprintf(
-				"Finished downloading %d GDrive files!",
-				len(allowedForDownload),
-			),
-			fmt.Sprintf(
-				"Something went wrong while downloading %d GDrive files!\nPlease refer to the generated log files for more details.",
-				len(allowedForDownload),
-			),
-			len(allowedForDownload),
-		)
-		progress.Start()
-		for _, file := range allowedForDownload {
-			wg.Add(1)
-			queue <- struct{}{}
-			go func(file map[string]string) {
-				defer wg.Done()
-
-				os.MkdirAll(file["filepath"], 0666)
-				filePath := filepath.Join(file["filepath"], file["name"])
-
-				if err := gdrive.DownloadFile(file, filePath); err != nil {
-					errorMsg := fmt.Sprintf(
-						"Failed to download file: %s (ID: %s, MIME Type: %s)\nRefer to error details below:\n%v",
-						file["name"], file["id"], file["mimeType"], err,
-					)
-					errChan <- map[string]string{
-						"err":	    errorMsg,
-						"filepath": filepath.Join(
-							file["filepath"], 
-							GDRIVE_ERROR_FILENAME,
-						),
-					}
-				}
-
-				progress.MsgIncrement(baseMsg)
-				<-queue
-			}(file)
-		}
-		close(queue)
-		wg.Wait()
-		close(errChan)
-
-		hasErr := false
-		if len(errChan) > 0 {
-			hasErr = true
-			for errMap := range errChan {
-				utils.LogMessageToPath(
-					errMap["err"],
-					errMap["filepath"], 
-				)
-			}
-		}
-		progress.Stop(hasErr)
-	}
-}
-
-// Uses regex to extract the file ID and the file type (type: file, folder) from the given URL
-func GetFileIdAndTypeFromUrl(url string) (string, string) {
-	matched := utils.GDRIVE_URL_REGEX.FindStringSubmatch(url)
-	if matched == nil {
-		return "", ""
-	}
-
-	var fileType string
-	matchedFileType := matched[utils.GDRIVE_REGEX_TYPE_INDEX]
-	if strings.Contains(matchedFileType, "folder") {
-		fileType = "folder"
-	} else if strings.Contains(matchedFileType, "file") {
-		fileType = "file"
-	} else {
-		err := fmt.Errorf(
-			"gdrive error %d: could not determine file type from URL, \"%s\"",
-			utils.DEV_ERROR,
-			url,
-		)
-		utils.LogError(err, "", false)
-		return "", ""
-	}
-	return matched[utils.GDRIVE_REGEX_ID_INDEX], fileType
-}
-
-// Downloads multiple GDrive files based on a slice of GDrive URL strings in parallel
-func (gdrive *GDrive) DownloadGdriveUrls(gdriveUrls []map[string]string) error {
-	if len(gdriveUrls) == 0 {
-		return nil
-	}
-
-	// Retrieve the id from the url text
-	gdriveIds := []map[string]string{}
-	for _, gdriveUrl := range gdriveUrls {
-		fileId, fileType := GetFileIdAndTypeFromUrl(gdriveUrl["url"])
-		if fileId != "" && fileType != "" {
-			gdriveIds = append(gdriveIds, map[string]string{
-				"id":       fileId,
-				"type":     fileType,
-				"filepath": gdriveUrl["filepath"],
-			})
-		}
-	}
-
-	var errMapSlice []map[string]string
-	gdriveFilesInfo := []map[string]string{}
-	baseMsg := "Getting gdrive file info from gdrive ID(s) [%d/" + fmt.Sprintf("%d]...", len(gdriveIds))
-	progress := spinner.New(
-		spinner.REQ_SPINNER,
-		"fgHiYellow",
-		fmt.Sprintf(
-			baseMsg,
-			0,
-		),
-		fmt.Sprintf(
-			"Finished getting gdrive file info from %d gdrive ID(s)!",
-			len(gdriveIds),
-		),
-		fmt.Sprintf(
-			"Something went wrong while getting gdrive file info from %d gdrive ID(s)!\nPlease refer to the generated log files for more details.",
-			len(gdriveIds),
-		),
-		len(gdriveIds),
-	)
-	progress.Start()
-	for _, gdriveId := range gdriveIds {
-		if gdriveId["type"] == "file" {
-			fileInfo, err := gdrive.GetFileDetails(gdriveId["id"], gdriveId["filepath"])
-			if err != nil {
-				errMapSlice = append(errMapSlice, map[string]string{
-					"err":	    err.Error(),
-					"filepath": gdriveId["filepath"],
-				})
-			} else if fileInfo != nil {
-				fileInfo["filepath"] = gdriveId["filepath"]
-				gdriveFilesInfo = append(gdriveFilesInfo, fileInfo)
-			}
-			progress.MsgIncrement(baseMsg)
-			continue
-		} 
-
-		if gdriveId["type"] == "folder" {
-			filesInfo, err := gdrive.GetNestedFolderContents(gdriveId["id"], gdriveId["filepath"])
-			if err != nil {
-				errMapSlice = append(errMapSlice, map[string]string{
-					"err":	    err.Error(),
-					"filepath": gdriveId["filepath"],
-				})
-			} else {
-				for _, fileInfo := range filesInfo {
-					fileInfo["filepath"] = gdriveId["filepath"]
-					gdriveFilesInfo = append(gdriveFilesInfo, fileInfo)
-				}
-			}
-			progress.MsgIncrement(baseMsg)
-			continue
-		} 
-
-		errMsg := fmt.Sprintf(
-			"gdrive error %d: unknown Google Drive URL type, \"%s\"",
-			utils.DEV_ERROR,
-			gdriveId["type"],
-		)
-		errMapSlice = append(errMapSlice, map[string]string{
-			"err":	    errMsg,
-			"filepath": gdriveId["filepath"],
-		})
-		progress.MsgIncrement(baseMsg)
-	}
-
-	hasErr := false
-	if len(errMapSlice) > 0 {
-		hasErr = true
-		for _, errMap := range errMapSlice {
-			utils.LogMessageToPath(
-				errMap["err"],
-				errMap["filepath"], 
-			)
-		}
-	}
-	progress.Stop(hasErr)
-
-	gdrive.DownloadMultipleFiles(gdriveFilesInfo)
-	return nil
 }
