+++ src/spinner/spinner_types.go
