+++ src/utils/constants.go
@@ -17,8 +17,8 @@ const (
 	INPUT_ERROR
 	CMD_ERROR
 	CONNECTION_ERROR
-	DOWNLOAD_ERROR
 	RESPONSE_ERROR
+	DOWNLOAD_ERROR
 	JSON_ERROR
 	HTML_ERROR
 )
